// src/utils/enums.ts
export enum AlertStatus {
  Info = "info",
  Warning = "warning",
  Success = "success",
  Error = "error",
  Loading = "loading",
}

export enum AlertColorScheme {
  WhiteAlpha = "whiteAlpha",
  BlackAlpha = "blackAlpha",
  Gray = "gray",
  Red = "red",
  Orange = "orange",
  Yellow = "yellow",
  Green = "green",
  Teal = "teal",
  Blue = "blue",
  Cyan = "cyan",
  Purple = "purple",
  Pink = "pink",
  Linkedin = "linkedin",
  Facebook = "facebook",
  Messenger = "messenger",
  Whatsapp = "whatsapp",
  Twitter = "twitter",
  Telegram = "telegram",
}